requirements.txt
amqp==1.4.9
anyjson==0.3.3
apiclient==1.0.4
appdirs==1.4.4
asgiref==3.2.10
attrs==20.3.0
Automat==20.2.0
billiard==3.3.0.23
cachetools==4.1.1
celery==3.1.26.post2
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.2.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.3
django-adminlte-ui==1.5.1
django-allauth==0.44.0
django-celery==3.3.1
django-cors-headers==3.2.1
django-createsuperuser==2020.12.3
django-elasticsearch-dsl==7.1.4
django-heroku==0.3.1
django-rest-knox==4.1.0
django-sslserver==0.22
django-treebeard==4.3
djangorestframework==3.12.1
elasticsearch==7.10.0
elasticsearch-dsl==7.3.0
filelock==3.0.12
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
kombu==3.0.37
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.4.0.46
pdf2image==1.14.0
Pillow==8.0.1
pipenv==2020.11.15
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2019.3
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.2.1
virtualenv-clone==0.5.4
whitenoise==5.2.0
zope.interface==4.7.1
